recipe databse refactor
