+++ mealie/db/models/_all_models.py
@@ -1,5 +1,5 @@
 from db.models.mealplan import *
-from db.models.recipe import *
+from db.models.recipe.recipe import *
 from db.models.settings import *
 from db.models.theme import *
 from db.models.users import *
