+++ mealie/services/backups/imports.py
@@ -6,11 +6,11 @@ from typing import List
 
 from core.config import BACKUP_DIR, IMG_DIR, TEMP_DIR
 from db.database import db
+from fastapi.logger import logger
 from schema.restore import RecipeImport, SettingsImport, ThemeImport
 from schema.theme import SiteTheme
 from services.recipe_services import Recipe
 from sqlalchemy.orm.session import Session
-from fastapi.logger import logger
 
 
 class ImportDatabase:
@@ -85,6 +85,11 @@ class ImportDatabase:
                 recipe_dict = json.loads(f.read())
                 recipe_dict = ImportDatabase._recipe_migration(recipe_dict)
             try:
+                if recipe_dict.get("categories", False):
+                    recipe_dict["recipeCategory"] = recipe_dict.get("categories")
+                    del recipe_dict["categories"]
+                    print(recipe_dict)
+
                 recipe_obj = Recipe(**recipe_dict)
                 recipe_obj.save_to_db(self.session)
                 import_status = RecipeImport(
